requirements.txt
mcp>=0.1.0
typing-extensions>=4.0.0
dataclasses>=0.6 
requests>=2.31.0
python-dotenv>=1.0.0 